chore: extract track types
